requirements.txt
anyio==4.4.0
APScheduler==3.10.4
argcomplete==3.5.0
attrs==24.2.0
bcrypt==4.2.0
black==24.8.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
datamodel-code-generator==0.25.6
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.115.2
fastapi-code-generator==0.5.1
genson==1.3.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-resources==6.4.3
inflect==5.6.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-spec==0.2.4
lazy-object-proxy==1.10.0
librouteros==3.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==24.1
paramiko==3.4.1
pathable==0.4.3
pathspec==0.12.1
platformdirs==4.2.2
prance==23.6.21.0
pycparser==2.22
pydantic==2.8.2
Pygments==2.18.0
PyNaCl==1.5.0
pyrsistent==0.20.0
PySnooper==1.1.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
shellingham==1.5.4
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
stringcase==1.2.0
swagger-ui-bundle==1.1.0
toml==0.10.2
typed-ast==1.5.5
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.6
Werkzeug==3.0.3
pylint==3.3.1
pylintfileheader==1.0.0
